// src/api/generated/model/listAddresses200Item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Nero API
 * A Nero API é uma interface RESTful robusta projetada especificamente para o varejo de moda online. Ela oferece controle total sobre o ciclo de vida do produto, desde a entrada de estoque em múltiplas variantes (SKUs) até o processamento final do checkout. Ideal para marketplaces, lojas boutique ou aplicativos móveis de moda.
 * OpenAPI spec version: 1.0.0
 */

export type ListAddresses200Item = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  userId: string;
  label: string | null;
  recipientName: string;
  street: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
  complement: string | null;
  isDefault: boolean;
  createdAt: string;
};